Refactoring of routes
