+++ lib/util.js
@@ -132,5 +132,9 @@ module.exports = {
     return function(req, res, next) {
       return accountService.hasRole(req.user.name, role).then(result => result ? next() : res.status(403).send())
     }
+  },
+
+  verifyMaintenanceToken:function(req, res, next) {
+    return req.query.verify === process.env['verifyToken'] ? next() : res.status(403).send();
   }
 };
