+++ routes/admin/strikes.js
